Fix downloads

+++ scripts/download_asistent_binarized.sh
@@ -4,9 +4,13 @@ DIR=data/data-binarized-asistent
 mkdir -p $DIR
 cd $DIR
 
-wget --no-check-certificate 'https://drive.google.com/uc?export=download&id=1XlN7mabhTT5W--bVhE7Up14gwQVkRV87' -O asistent_binarized.zip
+if [ -f data/data-binarized-asistent/dict.en.txt ]; then
 
-unzip asistent_binarized.zip
-rm asistent_binarized.zip
+  wget --no-check-certificate 'https://drive.google.com/uc?export=download&id=1XlN7mabhTT5W--bVhE7Up14gwQVkRV87' -O asistent_binarized.zip
+
+  unzip asistent_binarized.zip
+  rm asistent_binarized.zip
+
+fi
 
 cd ../..
